Added find by VisitorId

+++ src/main/java/com/cg/aps/service/VisitorService.java
@@ -19,7 +19,7 @@ public interface VisitorService {
 	
 	public List<VisitorEntity> findByName(String name);
 	
-	public Optional<VisitorEntity> findByPk(long id);
+	public Optional<VisitorEntity> findByPk(int id);
 	
 	public List<VisitorEntity> search(VisitorEntity bean, long pageNo, int pageSize);
